Move Label to outer Form and export the Choices components
