+++ lib/Form/Checkbox/Input.js
@@ -1,15 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const Input = props =>
+const Input = ({name, id, checked, disabled, onChangeHandler, ...rest}) =>
   <input
-    disabled={props.disabled}
+    disabled={disabled}
     className="form-checkbox__input"
-    type="checkbox"
-    name={props.name}
-    id={props.id}
-    onChange={props.onChangeHandler}
-    defaultChecked={props.checked}
+    type='checkbox'
+    name={name}
+    id={id}
+    onChange={onChangeHandler}
+    defaultChecked={checked}
+    {...rest}
   />;
 
 Input.propTypes = {
@@ -22,6 +23,7 @@ Input.propTypes = {
 
 Input.defaultProps = {
   checked: false,
+  name: '',
   disabled: false,
   onChangeHandler: () => {},
 };
