+++ lib/Form/Checkbox/index.js
@@ -1,4 +1,4 @@
-import Label from './Label';
+import Label from '../Label';
 import Input from './Input';
 
 const Checkbox = {
