+++ lib/Form/Label/index.js
