+++ lib/Form/RadioButton/Input.js
@@ -1,16 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const Input = (props) =>
+const Input = ({disabled, name, value, id, checked, onChangeHandler, ...rest}) =>
   <input
-    disabled={props.disabled}
+    disabled={disabled}
     className="form-radio__input"
+    name={name}
     type="radio"
-    name={props.name}
-    value={props.value}
-    id={props.id}
-    onChange={props.onChangeHandler}
-    defaultChecked={props.checked}
+    value={value}
+    id={id}
+    onChange={onChangeHandler}
+    defaultChecked={checked}
+    {...rest}
   />;
 
 Input.propTypes = {
@@ -24,6 +25,7 @@ Input.propTypes = {
 
 Input.defaultProps = {
   checked: false,
+  value: '',
   disabled: false,
   onChangeHandler: () => {},
 };
