+++ lib/Form/RadioButton/index.js
@@ -1,4 +1,4 @@
-import Label from '../Checkbox/Label';
+import Label from '../Label';
 import Input from './Input';
 
 const Radio = {
