+++ lib/index.js
@@ -20,3 +20,5 @@ export Modal from './Modal';
 export Field from './Field';
 export FileCard from './FileCard';
 export Conversation from './Conversation';
+export RadioButton from './Form/RadioButton';
+export Checkbox from './Form/Checkbox';
